+++ resources/views/sp2ds/surat_pengantar.blade.php
@@ -1,0 +1,8 @@
+@extends('layouts.admin')
+
+@section('title', __('Surat Pengantar'))
+
+@section('content')
+    <h2>Halaman Surat Pengantar</h2>
+    <p>Ini adalah halaman untuk mengelola Surat Pengantar.</p>
+@endsection